refactor: client.type 파일에 있던 server type들 server.type 파일로 옮김
