+++ types/client.types.ts
@@ -64,7 +64,7 @@ export type User = {
   userId: number;
   nickName: string;
   email: string;
-  gender: "female" | "male";
+  gender: "남자" | "여자";
   age: number;
 };
 
@@ -86,30 +86,6 @@ export type ImageType = {
 
 export type OrderValue = "인기순" | "평점순" | "최신순";
 
-export type UserAdditionalInfo = {
-  birthDate: string;
-  gender: string;
-  nickName: string;
-};
-
-export type UserSocialLoginData = {
-  id: number;
-  email: string;
-  nickName: string;
-  imageUrl: string;
-  role: string; //ROLE_USER, ROLE_GUEST
-  tokenType: string;
-  accessToken: string;
-};
-
-export type UserInfoType = {
-  email: string;
-  nickName: string;
-  imageUrl: string;
-  gender: string;
-  birthDate: string;
-  role: string;
-};
 export type birthdateValType = { value: number; label: number };
 export interface MapProps {
   locationIDList: string[];
